TextRun: only start/end required, size_pt u32 -> f32 (#47 part A)
format/text-runs.md's own opening example failed the typed loader
("missing field `italic`"), and CONCEPTS calls font/size_pt/color
optional while every one of them was required. The docs describe the
intended format; the struct now matches them: the three style flags,
`font`, `color` and `size_pt` all carry serde defaults chosen so an
omitted field renders exactly like the uncovered text around it —
false flags, unpinned font (empty, per fonts.md), deferred color
(empty, per text-runs.md's third-state rule), and a new
`DEFAULT_TEXT_RUN_SIZE_PT` (14) that `DEFAULT_SECTION_FONT_SCALE` is
now defined as, so "run that names no size" and "section with no
runs" cannot drift apart.

`size_pt` becomes `f32` because the schema types it "number" and it
was the only size field in the model that refused a fraction:
`GlyphBorderConfig.font_size_pt`, `GlyphConnectionConfig.font_size_pt`
and every label/portal clamp are f32, so `"size_pt": 14.5` parsed
everywhere except the one field users touch most. The integral field
also forced `.round()` calls and a checked f32->u32 narrowing through
the reverse converter; those are gone, fractional sizes are stored as
written, and validate.rs now runs run sizes through the same
`font_size` domain check (finite, 0.5pt floor, 4096pt ceiling) as
every sibling size instead of a private integer zero-check.

The saver omits fields still sitting at their defaults — the omission
policy schema.md already states — with both new predicates modeled in
OMITTABLE_WHEN so the round-trip key-loss tests stay strict.
`hyperlink` picks up the same `skip_serializing_if` (it was a second
Option field serializing `null` per run, alongside the edge-label one
part B removes). Fixtures re-saved in this commit per
CODE_CONVENTIONS §10; migration tooling (`maptool convert`) operates
on raw JSON for text runs and needed no change.

New loader-edge tests pin the minimal `{start, end}` shape, the
still-required range fields, authored values surviving the defaults,
save-side omission with its negative control, and text-runs.md's
example loading as written — read out of the spec via doc_fixtures,
which is #47's acceptance criterion for part A.

+++ lib/baumhard/src/mindmap/placard.rs
@@ -128,7 +128,7 @@ pub const PLACARD_HEADLINE: &str = "This map did not load.";
 pub const PLACARD_FOOTER: &str = "Nothing was opened; the file on disk is unchanged.";
 
 /// Point size of the placard's text run.
-const PLACARD_FONT_SIZE_PT: u32 = 24;
+const PLACARD_FONT_SIZE_PT: f32 = 24.0;
 
 /// Font family of the placard's text run: **empty, meaning no pin**.
 ///
